requirements.txt
google-genai>=0.3.0
google-generativeai>=0.4.0
chromadb>=0.4.22
python-telegram-bot[job-queue]>=21.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
SQLAlchemy>=2.0.0
openai>=1.0.0
PyMuPDF>=1.23.0
EbookLib>=0.18
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
langchain-text-splitters>=0.0.1
